Move blueprint-related into edit namespace.
