+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Edit/Blueprints/TextTagSelectionBlueprint.cs
@@ -10,12 +10,13 @@ using osu.Framework.Graphics.Primitives;
 using osu.Framework.Graphics.Shapes;
 using osu.Game.Graphics;
 using osu.Game.Rulesets.Edit;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
 using osu.Game.Rulesets.Karaoke.Objects.Types;
 using osu.Game.Rulesets.Karaoke.Utils;
 using osuTK;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.Blueprints
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Edit.Blueprints
 {
     public abstract class TextTagSelectionBlueprint<T> : SelectionBlueprint<T> where T : ITextTag
     {
