+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/Edit/Blueprints/TimeTagSelectionBlueprint.cs
@@ -6,10 +6,11 @@
 using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Game.Rulesets.Edit;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
 using osu.Game.Rulesets.Karaoke.Objects;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.Blueprints
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Edit.Blueprints
 {
     public class TimeTagSelectionBlueprint : SelectionBlueprint<TimeTag>
     {
